+++ tools/addstring.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-grep -RIl "\<string name=\"settings_dangerous" . | xargs sed -i -e '/string name="settings_dangerous/a \
-\ \ \ \ <string name="settings_confidence">Maximum fetch confidence interval</string>'
+grep -RIl "\<string name=\"menu_restrict_all" . | xargs sed -i -e '/string name="menu_restrict_all/a \
+\ \ \ \ <string name="menu_clear_usage">Clear</string>'
 
 #grep -RIl "\<string name=\"msg_edit" . | xargs sed -i -e '/msg_edit/d'